+++ x/auth/simulation/decoder.go
@@ -4,29 +4,39 @@ import (
 	"bytes"
 	"fmt"
 
+	gogotypes "github.com/gogo/protobuf/types"
 	tmkv "github.com/tendermint/tendermint/libs/kv"
 
-	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/x/auth/exported"
 	"github.com/cosmos/cosmos-sdk/x/auth/types"
 )
 
-// DecodeStore unmarshals the KVPair's Value to the corresponding auth type
-func DecodeStore(cdc *codec.Codec, kvA, kvB tmkv.Pair) string {
-	switch {
-	case bytes.Equal(kvA.Key[:1], types.AddressStoreKeyPrefix):
-		var accA, accB exported.Account
-		cdc.MustUnmarshalBinaryBare(kvA.Value, &accA)
-		cdc.MustUnmarshalBinaryBare(kvB.Value, &accB)
-		return fmt.Sprintf("%v\n%v", accA, accB)
+// NewDecodeStore returns a decoder function closure that unmarshals the KVPair's
+// Value to the corresponding auth type.
+func NewDecodeStore(cdc types.Codec) func(kvA, kvB tmkv.Pair) string {
+	return func(kvA, kvB tmkv.Pair) string {
+		switch {
+		case bytes.Equal(kvA.Key[:1], types.AddressStoreKeyPrefix):
+			accA, err := cdc.UnmarshalAccount(kvA.Value)
+			if err != nil {
+				panic(err)
+			}
 
-	case bytes.Equal(kvA.Key, types.GlobalAccountNumberKey):
-		var globalAccNumberA, globalAccNumberB uint64
-		cdc.MustUnmarshalBinaryBare(kvA.Value, &globalAccNumberA)
-		cdc.MustUnmarshalBinaryBare(kvB.Value, &globalAccNumberB)
-		return fmt.Sprintf("GlobalAccNumberA: %d\nGlobalAccNumberB: %d", globalAccNumberA, globalAccNumberB)
+			accB, err := cdc.UnmarshalAccount(kvB.Value)
+			if err != nil {
+				panic(err)
+			}
 
-	default:
-		panic(fmt.Sprintf("invalid account key %X", kvA.Key))
+			return fmt.Sprintf("%v\n%v", accA, accB)
+
+		case bytes.Equal(kvA.Key, types.GlobalAccountNumberKey):
+			var globalAccNumberA, globalAccNumberB gogotypes.UInt64Value
+			cdc.MustUnmarshalBinaryBare(kvA.Value, &globalAccNumberA)
+			cdc.MustUnmarshalBinaryBare(kvB.Value, &globalAccNumberB)
+
+			return fmt.Sprintf("GlobalAccNumberA: %d\nGlobalAccNumberB: %d", globalAccNumberA, globalAccNumberB)
+
+		default:
+			panic(fmt.Sprintf("unexpected %s key %X (%s)", types.ModuleName, kvA.Key, kvA.Key))
+		}
 	}
 }